app/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "livekit-agent"
version = "1.0.0"
description = "Single prompt voice AI agent built with LiveKit Agents for Python"
requires-python = ">=3.9"

dependencies = [
    "livekit-agents[silero,turn-detector,xai,openai]~=1.3",
    "livekit-plugins-noise-cancellation~=0.2",
    "python-dotenv",
    "python-handlebars>=0.0.3",
    "httpx",
    "pytz",
    "mem0ai"
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-asyncio",
    "ruff",
]

[tool.setuptools]
py-modules = ["agent"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "B", "A", "C4", "UP", "SIM", "RUF"]
ignore = ["E501", "W291", "RUF001"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
